0. Что такое тестирование_updated [#436736]

+++ src/ru/job4j/converter/Converter.java
@@ -6,15 +6,15 @@ public class Converter {
         return value / 70;
     }
 
-    public static float rubleToDollar(float value) {
+    public static int rubleToDollar(int value) {
         return value / 60;
     }
 
     public static void main(String[] args) {
         int expected1 = 2;
-        float expected2 = 3;
+        int expected2 = 3;
         int euro = Converter.rubleToEuro(140);
-        float dollar = Converter.rubleToDollar(180);
+        int dollar = Converter.rubleToDollar(180);
         boolean passedE = expected1 == euro;
         boolean passedD = expected2 == dollar;
         System.out.println("140 rubles are " + euro + " euro." + "Test result : " + passedE);
